Syntax fixed for OR operator (Balkan Wars)
In two event files related to the Balkan Wars, I discovered an OR operator that was not working correctly. In first example, it was used to allow Serbia to declare the First Balkan War if it was in an alliance with at least one country (Greece, Bulgaria or Montenegro). However, because all three countries were placed in one block within the OR operator rather than separated into three separate blocks, the OR operator worked as if there was only one condition with three attributes. This leads to the fact that if at least one country is not in the alliance by the time Serbia begins to perform the "Liberate The Balkans" focus, then Serbia will not be able to complete this focus, and the First Balkan War will not begin. I tested this with the example of Greece, when it refused to join the alliance by invitation, and the war did not start.