Fixed transfer of states after the end of the Balkan Wars
In the Second Balkan Wars file, I discovered that the transfer of states after the end of the war did not work correctly. Since Serbia is the head of the alliance, if it makes peace, then the other countries in the alliance also automatically make peace. However, since the condition for the transfer of states to other participating countries, firstly, is after Serbia makes peace, and secondly, requires that the country be involved in the war at the time of the transfer of states, it turns out that when Serbia makes peace and receives its new states; for other countries, the condition of being at war is not met.

It turns out that when a peace treaty with Bulgaria is signed, only Serbia receives its new states, but Greece, Montenegro and Turkey do not receive their new states (if they participated in the war).

To correct this problem, Serbia's peace is moved to the end of the event "Bulgaria Offers Peace" so that other countries can make peace earlier and be able to gain their new states.

In the First Balkan War file, the peace event is changed in a similar way.