Fixed "SER_first_balkan" focus
Serbia will not be able to complete the "SER_first_balkan" focus if Turkey returns the states it occupied to Greece. These are states 731, 184, 757, 756, 790, 164. This scenario is possible when playing as Turkey, when, in order to avoid the Balkan War, the player sacrifices part of the European territories (Thrace, Aegean Macedonia and the Aegean Islands), but the rest remains under his control (Albania, North Macedonia, East Montenegro, South Serbia and others).

This is possible because one of the conditions for completing the "SER_first_balkan" focus is the presence of state 184 (Thrace) as part of the Ottoman Empire. If you transfer it to Greece, the condition will no longer be met.

To fix this, need to change this condition. I propose to make the presence of at least one of the following states within the Ottoman Empire a necessary condition for the start of the Balkan War: 184 (Thrace), 760 (South Serbia), 761 (East Montenegro), 788 (East Rumelia). In this case, Turkey will not be able to avoid war, no matter which states it voluntarily returns.