Fixed "TUR_ww1_BalkanLeagueMenace" focus
The Ottoman Empire will not be able to complete the "TUR_ww1_BalkanLeagueMenace" focus if Bulgaria, Montenegro or Greece refuse to join the Balkan League. If the Ottoman Empire fails this focus, it will then be unable to complete the next focus "TUR_ww1_SecondBalkanWarOpportunity" and join the Second Balkan War.

This is possible because one of the conditions for completing the "TUR_ww1_BalkanLeagueMenace" focus is the presence of all of the above countries in the alliance. If at least one country refuses to join the Balkan League (even Montenegro), then the condition will no longer be met.

To fix this, you need to change this condition. A necessary condition for performing the trick, I propose to make the presence in the Balkan League of at least one of the above countries (plus Serbia as a founder), so that either Bulgaria is in an alliance with Serbia, or Montenegro, or Greece.

In addition, since this condition will now be met even if not all Balkan countries are included in the Balkan League, it makes sense to use the "hostile diplomatic relations" modifier only in relation to those countries that have entered into an alliance with Serbia. If Greece (or Montenegro) is not a member of the alliance, then it is somehow not logical for the Ottoman Empire to worsen diplomatic relations with it (Bulgaria is not on the list, relations with it are worsening in the previous focus, "TUR_ww1_AntiBulgaria").