Handling events related to unusual player behavior during the Balkan Wars
If you start the game in observer mode, the Balkan Wars will take place in historical mode: Serbia, Montenegro, Bulgaria and Greece will first attack the Ottoman Empire together (BW1), and then Serbia, Montenegro, Greece and the Ottoman Empire will fight with Bulgaria (BW2). But what if the player decides to add variety to the game and makes decisions that differ from the events of world history?

Events of the First Balkan War:

Example 1:
The player plays as the Ottoman Empire and decides to give Greece its national states: Macedonia and Thrace, as well as the Aegean Islands (the Dodecanese is excluded from the list, since there is a possibility that by the time the decision is made it will belong to Italy). In this case, the reason for the Greeks to fight the Turks disappears: all occupied national regions are transferred to Greece, and Central Epirus, which remains with the Ottoman Empire, is not a national region of Greece. It is logical that Greece will refuse war in this case. And, accordingly, there will be no reason to be in the Balkan League.

Example 2:
The player plays for Greece and refuses to join the Balkan League, or leaves it before the First Balkan War. In this case, after the end of the war, it will turn out that Central Epirus (state 758) and the western part of Aegean Macedonia (state 731) will be ownerless. In this case, it is proposed to give state 758 to the newly created Albania (as an alternative, Serbia can also be given, but definitely not to Bulgaria), and state 731 to Bulgaria.

Remark: in example 1, Greece is also not participating in the war, but only state 758 will remain ownerless, since state 731 was already transferred to Greece by the Ottoman Empire.

Example 3:
The player plays for Bulgaria and refuses to join the Balkan League, or leaves it before the First Balkan War. In this case, after the end of the war, it will turn out that the eastern part of Aegean Macedonia (state 184) and the eastern part of present-day North Macedonia (state 759) will be ownerless. In this case, it is proposed to give state 184 to Greece, and state 759 to Serbia (as was done only after the Second Balkan War).

Events of the Second Balkan War:

Example 4:
The player plays as Greece and does not participate in the First Balkan War (examples 1 and 2), or participates but withdraws from the Balkan League before the Second Balkan War. In this case, since Greece will not be a member of the Balkan League by the start of the Second Balkan War, it is logical not to invite Greece to the Anti-Bulgarian League, and then Greece will not participate in the Second Balkan War.

Not an example, but a fix:
According to historical events, the Aegean Islands (state 790) were transferred to Greece after the First Balkan War.